feat: AuthGuard 적용

+++ src/post/post.service.ts
@@ -7,6 +7,7 @@ import { CreatePostDto } from './dto/create-post.dto';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Post } from './post.entity';
 import { Repository } from 'typeorm';
+import { User } from '../auth/user.entity';
 
 @Injectable()
 export class PostService {
@@ -15,27 +16,29 @@ export class PostService {
     private postRepository: Repository<Post>,
   ) {}
 
-  async getPosts(page: number) {
+  async getPosts(page: number, user: User) {
     const perPage = 10;
     console.log(page);
     const offset = (page - 1) * perPage;
 
     return this.postRepository
       .createQueryBuilder('post')
+      .where('post.userId = :userId', { userId: user.id })
       .orderBy('post.date', 'DESC')
       .take(perPage)
       .skip(offset)
       .getMany();
   }
 
-  async getPostById(id: number) {
+  async getPostById(id: number, user: User) {
     try {
       const foundPost = await this.postRepository
         // post table에 있는 포스트
         .createQueryBuilder('post')
+        .where('post.userId = :userId', { userId: user.id })
         // 어떤 포스트냐면
         // where를 이용하면, post.id가 인자로 받는 id와 같은 id를 찾을 것.
-        .where('post.id = id', { id })
+        .andWhere('post.id = id', { id })
         // getMany가 아닌 getOne()
         .getOne();
 
@@ -51,7 +54,7 @@ export class PostService {
     }
   }
 
-  async createPost(createPostDto: CreatePostDto) {
+  async createPost(createPostDto: CreatePostDto, user: User) {
     const {
       latitude,
       longitude,
@@ -72,6 +75,7 @@ export class PostService {
       title,
       description,
       date,
+      user,
     });
 
     try {
@@ -83,17 +87,21 @@ export class PostService {
         '장소를 추가하는 도중 에러가 발생했습니다.',
       );
     }
-    return post;
+    // 유저 정보 빼고 넘기는법.
+    const { user: _, ...postWithoutUser } = post;
+    return postWithoutUser;
   }
 
-  async deletePost(id: number) {
+  async deletePost(id: number, user: User) {
     try {
       const result = await this.postRepository
         .createQueryBuilder('post')
         .delete()
         .from(Post)
+        // .where('post.userId = :userId', { userId: user.id })
+        .where('userId = :userId', { userId: user.id })
         // id가 인자와 똑같은 아이디를 삭제
-        .where('id = :id', { id })
+        .andWhere('id = :id', { id })
         .execute();
       // 결과가 일어나지 않으면
       if (result.affected === 0) {
@@ -111,8 +119,9 @@ export class PostService {
   async updatePost(
     id: number,
     updatePostDto: Omit<CreatePostDto, 'longitude' | 'latitude' | 'address'>,
+    user: User,
   ) {
-    const post = await this.getPostById(id);
+    const post = await this.getPostById(id, user);
 
     const { title, description, color, date, score, imageUris } = updatePostDto;
     post.title = title;
@@ -134,11 +143,12 @@ export class PostService {
     return post;
   }
 
-  async getAllMarkers() {
+  async getAllMarkers(user: User) {
     try {
       // 특정 칼럼만 갖어오는 Select
       const markers = await this.postRepository
         .createQueryBuilder('post')
+        .where('post.userId = :userId', { userId: user.id })
         .select([
           'post.id',
           'post.latitude',
